Rename dom-utils.ts to domUtils.ts
